bringing all BrokeredMessage tests under one roof
